+++ packages/keystone/lib/Keystone/index.js
@@ -3,6 +3,8 @@ const fs = require('fs');
 const gql = require('graphql-tag');
 const flattenDeep = require('lodash.flattendeep');
 const fastMemoize = require('fast-memoize');
+const falsey = require('falsey');
+const createCorsMiddleware = require('cors');
 const { print } = require('graphql/language/printer');
 const { graphql } = require('graphql');
 const {
@@ -12,10 +14,12 @@ const {
   objMerge,
   flatten,
   unique,
+  filterValues,
 } = require('@keystone-alpha/utils');
 const {
   validateFieldAccessControl,
   validateListAccessControl,
+  parseCustomAccess,
 } = require('@keystone-alpha/access-control');
 const {
   startAuthedSession,
@@ -50,7 +54,7 @@ module.exports = class Keystone {
   }) {
     this.name = name;
     this.adapterConnectOptions = adapterConnectOptions;
-    this.defaultAccess = { list: true, field: true, ...defaultAccess };
+    this.defaultAccess = { list: true, field: true, custom: true, ...defaultAccess };
     this.auth = {};
     this.lists = {};
     this.listsArray = [];
@@ -74,7 +78,7 @@ module.exports = class Keystone {
       this.adapters = { [adapter.constructor.name]: adapter };
       this.defaultAdapter = adapter.constructor.name;
     } else {
-      throw new Error('Need an adapter, yo');
+      throw new Error('No database adapter provided');
     }
 
     // Placeholder until keystone.prepare() is run during which this function
@@ -236,9 +240,18 @@ module.exports = class Keystone {
   }
 
   extendGraphQLSchema({ types = [], queries = [], mutations = [] }) {
-    this._extendedTypes = this._extendedTypes.concat(types);
-    this._extendedQueries = this._extendedQueries.concat(queries);
-    this._extendedMutations = this._extendedMutations.concat(mutations);
+    const _parseAccess = obj => ({
+      ...obj,
+      access: parseCustomAccess({
+        access: obj.access,
+        schemaNames: this._schemaNames,
+        defaultAccess: this.defaultAccess.custom,
+      }),
+    });
+
+    this._extendedTypes = this._extendedTypes.concat(types.map(_parseAccess));
+    this._extendedQueries = this._extendedQueries.concat(queries.map(_parseAccess));
+    this._extendedMutations = this._extendedMutations.concat(mutations.map(_parseAccess));
   }
 
   /**
@@ -290,6 +303,15 @@ module.exports = class Keystone {
     // not have any GraphQL operations performed on them
     const firstClassLists = this.listsArray.filter(list => !list.isAuxList);
 
+    const mutations = unique(
+      flatten([
+        ...firstClassLists.map(list => list.getGqlMutations({ schemaName })),
+        this._extendedMutations
+          .filter(({ access }) => access[schemaName])
+          .map(({ schema }) => schema),
+      ])
+    );
+
     // Fields can be represented multiple times within and between lists.
     // If a field defines a `getGqlAuxTypes()` method, it will be
     // duplicated.
@@ -297,7 +319,9 @@ module.exports = class Keystone {
     // Deduping here avoids that problem.
     return [
       ...unique(flatten(this.listsArray.map(list => list.getGqlTypes({ schemaName })))),
-      ...unique(this._extendedTypes),
+      ...unique(
+        this._extendedTypes.filter(({ access }) => access[schemaName]).map(({ type }) => type)
+      ),
       `"""NOTE: Can be JSON, or a Boolean/Int/String
           Why not a union? GraphQL doesn't support a union including a scalar
           (https://github.com/facebook/graphql/issues/215)"""
@@ -321,7 +345,11 @@ module.exports = class Keystone {
           """Access Control settings for the currently logged in (or anonymous)
              user when performing 'delete' operations."""
           delete: JSON
-       }`,
+
+          """Access Control settings for the currently logged in (or anonymous)
+             user when performing 'auth' operations."""
+          auth: JSON
+        }`,
       `type _ListSchemaRelatedFields {
         """The typename as used in GraphQL queries"""
         type: String
@@ -359,21 +387,21 @@ module.exports = class Keystone {
           ${unique(
             flatten([
               ...firstClassLists.map(list => list.getGqlQueries({ schemaName })),
-              this._extendedQueries.map(({ schema }) => schema),
+              this._extendedQueries
+                .filter(({ access }) => access[schemaName])
+                .map(({ schema }) => schema),
             ])
           ).join('\n')}
           """ Retrieve the meta-data for all lists. """
           _ksListsMeta: [_ListMeta]
        }`,
-      `type Mutation {
-          ${unique(
-            flatten([
-              ...firstClassLists.map(list => list.getGqlMutations({ schemaName })),
-              this._extendedMutations.map(({ schema }) => schema),
-            ])
-          ).join('\n')}
+      mutations.length > 0 &&
+        `type Mutation {
+          ${mutations.join('\n')}
        }`,
-    ].map(s => print(gql(s)));
+    ]
+      .filter(s => s)
+      .map(s => print(gql(s)));
   }
 
   // It's not Keystone core's responsibility to create an executable schema, but
@@ -408,6 +436,7 @@ module.exports = class Keystone {
       read: access => access.getRead(),
       update: access => access.getUpdate(),
       delete: access => access.getDelete(),
+      auth: access => access.getAuth(),
     };
 
     // Aux lists are only there for typing and internal operations, they should
@@ -449,38 +478,45 @@ module.exports = class Keystone {
           resolver(obj, args, context, info, { query: this._buildQueryHelper(context) }),
       };
     };
-    const resolvers = {
-      // Order of spreading is important here - we don't want user-defined types
-      // to accidentally override important things like `Query`.
-      ...objMerge(this.listsArray.map(list => list.gqlAuxFieldResolvers({ schemaName }))),
-      ...objMerge(this.listsArray.map(list => list.gqlFieldResolvers({ schemaName }))),
+    const resolvers = filterValues(
+      {
+        // Order of spreading is important here - we don't want user-defined types
+        // to accidentally override important things like `Query`.
+        ...objMerge(this.listsArray.map(list => list.gqlAuxFieldResolvers({ schemaName }))),
+        ...objMerge(this.listsArray.map(list => list.gqlFieldResolvers({ schemaName }))),
 
-      JSON: GraphQLJSON,
+        JSON: GraphQLJSON,
 
-      _QueryMeta: queryMetaResolver,
-      _ListMeta: listMetaResolver,
-      _ListAccess: listAccessResolver,
-      _ListSchema: listSchemaResolver,
+        _QueryMeta: queryMetaResolver,
+        _ListMeta: listMetaResolver,
+        _ListAccess: listAccessResolver,
+        _ListSchema: listSchemaResolver,
 
-      Query: {
-        // Order is also important here, any TypeQuery's defined by types
-        // shouldn't be able to override list-level queries
-        ...objMerge(firstClassLists.map(list => list.gqlAuxQueryResolvers())),
-        ...objMerge(firstClassLists.map(list => list.gqlQueryResolvers({ schemaName }))),
-        // And the Keystone meta queries must always be available
-        _ksListsMeta: (_, args, context) =>
-          this.listsArray
-            .filter(list => list.access[schemaName].read)
-            .map(list => list.listMeta(context)),
-        ...objMerge(this._extendedQueries.map(customResolver)),
+        Query: {
+          // Order is also important here, any TypeQuery's defined by types
+          // shouldn't be able to override list-level queries
+          ...objMerge(firstClassLists.map(list => list.gqlAuxQueryResolvers())),
+          ...objMerge(firstClassLists.map(list => list.gqlQueryResolvers({ schemaName }))),
+          // And the Keystone meta queries must always be available
+          _ksListsMeta: (_, args, context) =>
+            this.listsArray
+              .filter(list => list.access[schemaName].read)
+              .map(list => list.listMeta(context)),
+          ...objMerge(
+            this._extendedQueries.filter(({ access }) => access[schemaName]).map(customResolver)
+          ),
+        },
+
+        Mutation: {
+          ...objMerge(firstClassLists.map(list => list.gqlAuxMutationResolvers())),
+          ...objMerge(firstClassLists.map(list => list.gqlMutationResolvers({ schemaName }))),
+          ...objMerge(
+            this._extendedMutations.filter(({ access }) => access[schemaName]).map(customResolver)
+          ),
+        },
       },
-
-      Mutation: {
-        ...objMerge(firstClassLists.map(list => list.gqlAuxMutationResolvers())),
-        ...objMerge(firstClassLists.map(list => list.gqlMutationResolvers({ schemaName }))),
-        ...objMerge(this._extendedMutations.map(customResolver)),
-      },
-    };
+      o => Object.entries(o).length > 0
+    );
 
     if (debugGraphQLSchemas()) {
       console.log(resolvers);
@@ -543,7 +579,13 @@ module.exports = class Keystone {
     return mergeRelationships(createdItems, createdRelationships);
   }
 
-  async prepare({ dev = false, apps = [], distDir } = {}) {
+  async prepare({
+    dev = false,
+    apps = [],
+    distDir,
+    pinoOptions,
+    cors = { origin: true, credentials: true },
+  } = {}) {
     const middlewares = flattenDeep([
       // Used by other middlewares such as authentication strategies. Important
       // to be first so the methods added to `req` are available further down
@@ -555,6 +597,8 @@ module.exports = class Keystone {
         secureCookies: this._secureCookies,
         cookieMaxAge: this._cookieMaxAge,
       }),
+      falsey(process.env.DISABLE_LOGGING) && require('express-pino-logger')(pinoOptions),
+      cors && createCorsMiddleware(cors),
       ...(await Promise.all(
         [
           // Inject any field middlewares (eg; WYSIWIG's static assets)